pyproject.toml
[tool.poetry]
name = "drova_keenetic_desktop"
version = "0.1.0"
description = ""
authors = ["sergius-dart <sergius-dart@yandex.ru>"]
readme = "README.md"
packages = [{include = "drova_desktop_keenetic"}]

[tool.poetry.scripts]
drova_validate = "drova_desktop_keenetic.bin.drova_validate:main"
drova_socket = "drova_desktop_keenetic.bin.drova_socket:main"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.1"
mslex = "^1.3.0"
pydantic = "^2.11.7"
aiohttp = "^3.12.15"
asyncio = "^4.0.0"
aiofiles = "^24.1.0"
asyncssh = "^2.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"
types-aiofiles = "^24.1.0.20250822"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://black.readthedocs.io/en/stable/usage_and_configuration/the_basics.html#command-line-options
line-length = 120
target-version = ['py311']

[tool.pytest.ini_options]
pythonpath = ["drova_desktop_keenetic"]

[tool.mypy]
exclude = [
    "dist"
]

[tool.isort]
profile = "black"